chore: move components to correct folder
