+++ packages/desktop/components/menus/index.ts
@@ -7,6 +7,5 @@ export { default as DappActionsMenu } from './DappActionsMenu.svelte'
 export { default as DappListActionsMenu } from './DappListActionsMenu.svelte'
 export { default as ProfileActionsMenu } from './ProfileActionsMenu.svelte'
 export { default as NodeActionsMenu } from './NodeActionsMenu.svelte'
-export { default as ProposalDetailsMenu } from './ProposalDetailsMenu.svelte'
 export { default as TokenActionsMenu } from './TokenActionsMenu.svelte'
 export { default as TokenListMenu } from './TokenListMenu.svelte'
