+++ packages/desktop/views/dashboard/governance/components/ProposalDetailsMenu.svelte
