+++ packages/desktop/views/dashboard/governance/components/ProposalInformationPane.svelte
