+++ packages/desktop/views/dashboard/governance/components/ProposalListDetails.svelte
