+++ packages/desktop/views/dashboard/governance/components/ProposalStatusPill.svelte
