+++ packages/desktop/views/dashboard/governance/components/ProposalStatusTimelineTooltip.svelte
