+++ packages/desktop/views/dashboard/governance/components/index.ts
@@ -1,2 +1,12 @@
 export { default as ManageVotingPowerMenu } from './ManageVotingPowerMenu.svelte'
 export { default as ManageVotingPowerPane } from './ManageVotingPowerPane.svelte'
+export { default as ProposalAnswer } from './ProposalAnswer.svelte'
+export { default as ProposalCard } from './ProposalCard.svelte'
+export { default as ProposalDetailsMenu } from './ProposalDetailsMenu.svelte'
+export { default as ProposalInformationPane } from './ProposalInformationPane.svelte'
+export { default as ProposalList } from './ProposalList.svelte'
+export { default as ProposalListDetails } from './ProposalListDetails.svelte'
+export { default as ProposalQuestion } from './ProposalQuestion.svelte'
+export { default as ProposalStatusInfo } from './ProposalStatusInfo.svelte'
+export { default as ProposalStatusPill } from './ProposalStatusPill.svelte'
+export { default as ProposalStatusTimelineTooltip } from './ProposalStatusTimelineTooltip.svelte'
