--- packages/desktop/views/governance/components/index.ts
@@ -1,6 +1,0 @@
-export { default as ProposalInformationPane } from './ProposalInformationPane.svelte'
-export { default as ProposalList } from './ProposalList.svelte'
-export { default as ProposalQuestion } from './ProposalQuestion.svelte'
-export { default as ProposalStatusInfo } from './ProposalStatusInfo.svelte'
-export { default as ProposalStatusPill } from './ProposalStatusPill.svelte'
-export { default as ProposalStatusTimelineTooltip } from './ProposalStatusTimelineTooltip.svelte'
